move tool to separate package; refactor serde
