+++ haystack/tools/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-FileCopyrightText: 2022-present deepset GmbH <info@deepset.ai>
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from haystack.tools.tool import Tool, _check_duplicate_tool_names, deserialize_tools_inplace
+
+__all__ = ["Tool", "_check_duplicate_tool_names", "deserialize_tools_inplace"]
